Rename service to `BillRunsService`
As we started to look at the next stage we realised that this should be called `BillRunsService` rather than `SystemService`, since each repo's connector has multiple services, each relating to a specific thing; in this case, working with bill runs.
